signiling/requirements.txt
aiohttp==3.8.1
python-socketio==5.5.1
python-engineio==4.3.1